chat: now using request_parse function in chat_request_prepare function
using request_parse function in chat_request_prepare function instead of
old way to split command message and then finding each parts of the
command which was inefficient and lead to complex and buggy code.
Now it is good and code is more simple for handling request. For this
reason some code has been changed in chat_request_prepare function.