chat: bug: fixed a bug, was freeing unfreeable memory of irc_cmd
In the previous commit server was needed to free irc_cmd member of
request struct but client was also using this request struct to set
irc_cmd and in the client side client was trying to free the irc_cmd
pointer but the pointer was a static const string which is not freeable
therefore causing abort in client app. Now fixed
Now irc_cmd is freed in chat_request_prepare function in server.